organize sync services
