--- backend/src/Designer/Middleware/UserRequestSynchronization/IRequestSyncEvaluator.cs
@@ -1,8 +1,0 @@
-using Microsoft.AspNetCore.Http;
-
-namespace Altinn.Studio.Designer.Middleware.UserRequestSynchronization;
-
-public interface IRequestSyncEvaluator
-{
-    bool EvaluateSyncRequest(HttpContext httpContext);
-}
