+++ backend/src/Designer/Middleware/UserRequestSynchronization/Services/IRequestSyncEvaluator.cs
@@ -1,0 +1,8 @@
+using Microsoft.AspNetCore.Http;
+
+namespace Altinn.Studio.Designer.Middleware.UserRequestSynchronization.Services;
+
+public interface IRequestSyncEvaluator
+{
+    bool IsEligibleForSynchronization(HttpContext httpContext);
+}
