+++ backend/src/Designer/Middleware/UserRequestSynchronization/Services/IRequestSyncResolver.cs
@@ -1,7 +1,7 @@
 using Altinn.Studio.Designer.Models;
 using Microsoft.AspNetCore.Http;
 
-namespace Altinn.Studio.Designer.Middleware.UserRequestSynchronization;
+namespace Altinn.Studio.Designer.Middleware.UserRequestSynchronization.Services;
 
 public interface IRequestSyncResolver
 {
